Sprint 15 hw (clean arc)
