--- app/src/main/java/com/practicum/playlistmaker_1/TrackResponse.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmaker_1
-
-class TrackResponse(
-    val results: List<Track>
-)
