+++ app/src/main/java/com/practicum/playlistmaker_1/data/TrackResponse.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker_1.data
+
+import com.practicum.playlistmaker_1.domain.models.Track
+
+class TrackResponse(
+    val results: List<Track>
+)
