+++ app/src/main/java/com/practicum/playlistmaker_1/domain/enums/PlayerState.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker_1
+package com.practicum.playlistmaker_1.domain.enums
 
 enum class PlayerState {
     STATE_DEFAULT,
