+++ app/src/main/java/com/practicum/playlistmaker_1/presentation/search/TrackViewHolder.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker_1
+package com.practicum.playlistmaker_1.presentation.search
 
 import android.view.View
 import android.widget.ImageView
@@ -6,6 +6,8 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.practicum.playlistmaker_1.R
+import com.practicum.playlistmaker_1.domain.models.Track
 import java.text.SimpleDateFormat
 import java.util.*
